refactor: 캐러셀 구조 분리 및 훅 분리
