+++ packages/shared/src/hooks/index.ts
@@ -1,1 +1,2 @@
+export { useDraggableCarousel } from './use-draggable-carousel';
 export { useNoticeList } from './use-notice-list';
